refactor(notifier): makes details modal reusable
