+++ src/components/organisms/notifier/mypurchase/details.ts
@@ -1,0 +1,10 @@
+import { baseSubscriptionHeaders } from '../details/NotifierDetailsModal'
+
+export const subscriptionHeaders = {
+  ...baseSubscriptionHeaders,
+  provider: 'Provider',
+}
+
+export type SubscriptionDetails = {
+    [K in keyof typeof subscriptionHeaders]: string
+  }
